Move server classes out of client
